refactor: ProductFormStyle 리팩토링
